typegen utils
